Add user collections API with JWT auth middleware
Implemented JWT-based authentication middleware and new endpoints for saving movies to user collections (favorites, watchlist) in the backend. Added corresponding API routes and context handling. Updated Movie model to use float64 for score and popularity. Introduced new frontend components and services to support account pages and state management.

+++ models/movie.go
@@ -8,8 +8,8 @@ type Movie struct {
 	ReleaseYear int      `json:"release_year"`
 	Genres      []Genre  `json:"genres"`
 	Overview    *string  `json:"overview,omitempty"`
-	Score       *float32 `json:"score,omitempty"`
-	Popularity  *float32 `json:"popularity,omitempty"`
+    Score       *float64 `json:"score,omitempty"`
+    Popularity  *float64 `json:"popularity,omitempty"`
 	Keywords    []string `json:"keywords"`
 	Language    *string  `json:"language,omitempty"`
 	PosterUrl   *string  `json:"poster_url,omitempty"`
